feat: 채팅방 페이지 동적 파라미터 이름 변경 및 채팅방 상세 조회 fetch (#202)
